Null-terminate attachment filename
It could be argued that it is the caller's responsibility to check the
size member, and not assume the string is null-terminated. But since
we are allocating memory anyway, might as well add an extra byte since
API users tend to expect stuff like filenames to be null-terminated.

It is already done like this in decompressRTF, but perhaps other places
could use this change as well